+++ docs/analytics/changelog.md
@@ -23,6 +23,8 @@ Added:
     -   autodetectTheme: boolean;
 -   suite-ready
     -   customBackends: ['btc', 'eth', 'ada', ...] (array of coins with custom backend)
+    -   autodetectLanguage: boolean;
+    -   autodetectTheme: boolean;
 -   settings/coin-backend
     -   symbol: 'btc', 'eth', 'ada', ...
     -   type: 'default' | 'blockbook' | 'electrum' | 'ripple' | 'blockfrost'